+++ gax-java/pom.xml
@@ -205,6 +205,11 @@
       </exclusions>
       <scope>test</scope>
     </dependency>
+    <dependency>
+      <groupId>uk.org.webcompere</groupId>
+      <artifactId>system-stubs-jupiter</artifactId>
+      <version>2.1.6</version>
+    </dependency>
   </dependencies>
 
   <build>
